Ahorn is a bit faster now.
Reduced amount of file access.
Improved mod loading system.
Made mod loading faster.
Made graphics loading faster.
Made reading zip files faster.
Made autotiler faster.
Made tile rendering faster.
Made redrawing just about anything faster.
Made entity and trigger rendering faster.
Made spike rendering faster.
Made cobweb rendering faster.
Made cassette block rendering faster.
Made text rendering faster.
Made the RNG faster.
Made selections faster.
Made sprite rendering faster.
Reduced lag on first save.
Reduced lag after initial load.
Removed and then re-added room h-02 from chapter 9.

For plugin developers:
Entity rendering with layer as first argument has been removed.

+++ src/camera.jl
@@ -7,7 +7,7 @@ mutable struct Camera
     Camera(x::Number, y::Number, scale::Number) = new(x, y, scale, false)
 end
 
-function setCamera!(camera::Camera, x::Integer, y::Integer)
+function setCamera!(camera::Camera, x, y)
     if !camera.locked
         camera.x = x * camera.scale
         camera.y = y * camera.scale
